Fix #1.
The error was causing an exception to bypass the popping of
the record scope, and we were deleting into the inside of
a record-decl subobject.

Extend the scope sentinel to more gracefully handle cases where
a pre-existing scope is pushed onto the stack. Use this feature
for managing the scope of record declarations.